chore: ci.yml 더미데이터 추가 chore(auth): CustomOauth2UserServiceTest 삭제 refactor(auth): codeStyle 수정 refactor(auth): codeStyle 수정 Merge branch 'dev' into feat/13/JJY refactor(member, auth): member도메인의 logout을 auth도메인으로 이동, 그 밖에 리뷰 반영 feat(member): UpdateMemberRequest DTO 추가 및 회원 수정 요청 적용 test(member, auth): logout member에서 auth로 이동 테스트 수정
